refactor: getRestaurants 함수 Context에서 분리
getRestaurants 함수를 RestaurantContext에서 삭제한 후, RestuarantList 컴포넌트에 위치시켰습니다.
