+++ src/context/ModalContext.jsx
@@ -3,7 +3,8 @@ import { createContext, useContext, useState } from "react";
 const ModalContext = createContext();
 
 export function ModalProvider({ children }) {
-  const [modalState, setModalState] = useState("null"); // detail, add, null 3가지로 관리
+  // detail, add, add-success, null 4가지로 관리
+  const [modalState, setModalState] = useState("null");
 
   return (
     <ModalContext.Provider
