+++ src/context/RestaurantContext.jsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, useState, useEffect } from "react";
+import { createContext, useContext, useState } from "react";
 
 const RestaurantContext = createContext();
 
@@ -9,20 +9,6 @@ export function RestaurantProvider({ children }) {
     description: "",
   });
 
-  const getRestaurants = async () => {
-    try {
-      const response = await fetch("http://localhost:3000/restaurants");
-      const data = await response.json();
-      setRestaurants(data);
-    } catch (err) {
-      console.error("레스토랑 데이터를 불러오는 데 실패했습니다:", err);
-    }
-  };
-
-  useEffect(() => {
-    getRestaurants();
-  }, []);
-
   return (
     <RestaurantContext.Provider
       value={{
@@ -30,7 +16,6 @@ export function RestaurantProvider({ children }) {
         setRestaurants,
         restaurantItem,
         setRestaurantItem,
-        getRestaurants,
       }}
     >
       {children}
